Add some error handling
Here I have used if let syntax, copying code from the book. I am not
sure this is the best example of if let usage though. If let could make
sense if you want to ignore some branches of your pattern matching, but
as soon as you need to add an else branch you might as well use an
explicit match for it.