Add answers data structure.
In the book thei're plying with form encoded data. At the moment I am
not interested in experimenting with that, but I want to have the
structures in place so I can follow along easily with the next examples.
I could implement json version of the api for the answers too, but it
would be a boring repetition.